requirements.txt
streamlit==1.38.0
torch==2.2.2
transformers==4.40.1
pandas==2.2.2
google-generativeai==0.8.0
safetensors==0.4.3